Revert "Add model user to models.py add migration "add_user_table""
This reverts commit 8dfbb7683f66e88a49bdf1d582cb87be6050eb09.
