--- alembic/versions/b831723808df_add_user_table.py
@@ -1,54 +1,0 @@
-"""add_user_table
-
-Revision ID: b831723808df
-Revises: 
-Create Date: 2023-09-27 11:50:36.057806
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b831723808df'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('user_email', sa.String(), nullable=True),
-    sa.Column('user_firstname', sa.String(), nullable=True),
-    sa.Column('user_lastname', sa.String(), nullable=True),
-    sa.Column('birthday', sa.DateTime(), nullable=True),
-    sa.Column('user_status', sa.Boolean(), nullable=True),
-    sa.Column('user_city', sa.String(), nullable=True),
-    sa.Column('user_phone', sa.String(), nullable=True),
-    sa.Column('user_links', sa.String(), nullable=True),
-    sa.Column('user_avatar', sa.String(), nullable=True),
-    sa.Column('hashed_password', sa.String(), nullable=True),
-    sa.Column('is_superuser', sa.Boolean(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('user_id')
-    )
-    op.create_index(op.f('ix_users_created_at'), 'users', ['created_at'], unique=False)
-    op.create_index(op.f('ix_users_updated_at'), 'users', ['updated_at'], unique=False)
-    op.create_index(op.f('ix_users_user_email'), 'users', ['user_email'], unique=True)
-    op.create_index(op.f('ix_users_user_id'), 'users', ['user_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_users_user_id'), table_name='users')
-    op.drop_index(op.f('ix_users_user_email'), table_name='users')
-    op.drop_index(op.f('ix_users_updated_at'), table_name='users')
-    op.drop_index(op.f('ix_users_created_at'), table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
