+++ app/core/config.py
@@ -31,3 +31,4 @@ class Settings:
     REDIS_URL = f"redis://{REDIS_HOST}:{REDIS_PORT}"
     ACCESS_TOKEN_EXPIRY_TIME = 60 * 30
     REFRESH_TOKEN_EXPIRY_TIME = 60 * 24 * 365
+
