+++ app/routers/user_routers.py
@@ -1,24 +1,17 @@
-from fastapi import HTTPException, Depends
-from sqlalchemy import text
-from sqlalchemy.ext.asyncio import AsyncSession
-from starlette.responses import JSONResponse
-from app.db.db import get_db
+from fastapi import HTTPException
+from app.db.db import engine
+from app.db.db import User
 from health import router
 
 
-# @router.get("/check_user_table")
-# async def check_user_table(session: AsyncSession = Depends(get_db)):
-#     try:
-#         result = await session.execute(text("SELECT 1 FROM users LIMIT 1"))
-#
-#         if result.scalar() == 1:
-#             return JSONResponse(content={"status_code": 200, "detail": "User table exist", "result": "working"})
-#         else:
-#             raise HTTPException(status_code=500, detail="User table do not exist")
-#
-#     except Exception as e:
-#
-#         raise HTTPException(status_code=500, detail=f"Error checking PostgreSQL connection: {e}")
+@router.get("/check_user_table")
+async def check_user_table():
+    try:
+        User.metadata.create_all(bind=engine)
+        return {"message": "User table exists"}
+
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Error checking User table: {e}")
 
 
 @router.get("/user_get")
